+++ NEWS.md
@@ -1,3 +1,9 @@
+# ecoevosimulator 1.0.9007
+* fixed sinusoidal topography generator and added Paracou topography
+
+# ecoevosimulator 1.0.9006
+* topography generators
+
 # ecoevosimulator 1.0.9005
 * fixed indexing issue on border for forestgapdynamics
 * renamed parameters following bookdown document